X Goodbye xio.Name()
It was not a good idea that objects should always have name - some don't
do e.g. bytes.Reader (analog of StringIO.StringIO in python).

Don't try to automatically unwrap IO wrappers - just check in fs1
directly - if object has Name - it will be included in decode/check
errors as prefix; if not - not included.

LimitedWriter was not used anywhere so also killed.