X neo/storage: Open backends by raw string, not URL
For both current backends (SQLite and fs1) there is currently a problem
of working with UTF-8 filenames if opening is handled via urls. For
example if backend string is

	sqlite://neo.sqlite·P1-1

after parsing as URL and further reassembled as string it will be

	neo.sqlite%C2%B7P1-1

with then errors like

	neo.sqlite%C2%B7P1-1: open: unable to open database file

showing.

Since storage backends primarily work with local files it makes sense
not to play raw url reassembly games and just pass raw strings to
backends for open.

SQLite uri processing is still possible, via e.g.

	sqlite://file:filepath?param=value,...

See

	https://www.sqlite.org/c3ref/open.html	and
	https://www.sqlite.org/uri.html

for details.