storage: also commit updated cell TID at each replicated chunk of 'obj' records
This is a follow-up of commit b3dd69730cf0e4273e1be33ee3a5ee382836b3b3
("Optimize resumption of replication by starting from a greater TID").
I missed the case where a storage node is restarted while it is replicating:
it lost the TID where it was interrupted.

Although we commit after each replicated chunk, to avoid transferring again
all the data from the beginning, it could still waste time to check that
the data are already replicated.