go/zodb/zeo: Accept both int|bool for pickled message flags
It is already documented in pktDecodeZ comment that flags is int|bool.
However until now we were decoding it only as int.

As is it was working, but it will fail when receiving e.g.
invalidateTransaction message, because ZEO/py server actually uses bool
when sending it:

    https://github.com/zopefoundation/ZEO/blob/5.2.1-20-gcb26281d/src/ZEO/asyncio/base.py#L139

Fix it.
This will be covered by watch tests, when watch support will be added in a followup patch.